+++ engine/src/game/game_log.cpp
@@ -9,7 +9,7 @@ GameLog GameLogStorage::view() const {
   v.initial_scores = initial_scores;
   v.initial_racks = initial_racks;
   v.records = turns.data();
-  v.num_records = static_cast<int>(turns.size());
+  v.num_records = turns.size();
   v.final_scores = final_scores;
   v.final_racks = final_racks;
   v.end_reason = end_reason.c_str();
